Remove adapter by enhancing sympy_utils
