+++ apps/prairielearn/elements/pl-big-operator-input/pl-big-operator-input_test.py
@@ -951,12 +951,11 @@ class TestLifecycleRegressions:
         assert big_operator_input._identifier(token) is None
 
 
-class TestAdapterRegressions:
+class TestSymbolicInputRendering:
     def test_score_badges_do_not_show_percent(self) -> None:
         data = question_data(raw_submitted_answers={"op": "x"}, panel="submission")
         data["submitted_answers"] = {"op": "x"}
-        adapter = big_operator_input.symbolic_input_adapter
-        s = adapter.render(
+        rendered = big_operator_input._render_symbolic_input(
             data,
             name="op",
             variables=("x",),
@@ -968,5 +967,5 @@ class TestAdapterRegressions:
             show_score=True,
             score=0.5,
         )
-        assert "text-bg-warning" in s
-        assert re.search(r"\d+(?:\.\d+)?%", s) is None
+        assert "text-bg-warning" in rendered
+        assert re.search(r"\d+(?:\.\d+)?%", rendered) is None
