+++ apps/prairielearn/python/pyproject.toml
@@ -14,6 +14,7 @@ readme = "README.md"
 requires-python = ">=3.13,<3.14"
 classifiers = ["Private :: Do Not Upload"]
 dependencies = [
+  "chevron==0.14.0",
   "coloraide==8.10",
   "jsonschema==4.26.0",
   "lxml==6.1.1",
